# app/api/routes/thoughts.py
import logging
from datetime import datetime
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Query, Response, status

from app.api.dependencies import CurrentUser, DbSession, IdempotencyKey
from app.core.config import settings
from app.models import SourceType, ThoughtType
from app.schemas import ThoughtCreate, ThoughtRead, ThoughtUpdate
from app.services.data_lifecycle import list_deleted_thoughts, restore_thought
from app.services.idempotency import execute_idempotent
from app.services.openai_ai import OpenAIProvider
from app.services.thought_recall import RecallQuery, list_thoughts
from app.services.thoughts import (
    create_thought,
    get_thought,
    retry_ai_processing,
    soft_delete_thought,
    update_thought,
)

router = APIRouter(prefix="/thoughts", tags=["thoughts"])
logger = logging.getLogger(__name__)


@router.post("", response_model=ThoughtRead, status_code=status.HTTP_201_CREATED)
def create_thought_route(
    payload: ThoughtCreate,
    db: DbSession,
    user: CurrentUser,
    response: Response,
    idempotency_key: IdempotencyKey = None,
):
    return execute_idempotent(
        db, user, "POST /thoughts", idempotency_key, payload.model_dump(mode="json"), response,
        lambda operation: create_thought(db, user, payload, operation=operation),
    )


@router.get("", response_model=list[ThoughtRead])
def list_thoughts_route(
    db: DbSession,
    user: CurrentUser,
    response: Response,
    q: Annotated[str | None, Query(max_length=200)] = None,
    thought_type: ThoughtType | None = None,
    source_type: SourceType | None = None,
    book_id: UUID | None = None,
    tag: Annotated[str | None, Query(max_length=100)] = None,
    book: Annotated[str | None, Query(max_length=255)] = None,
    theme: Annotated[str | None, Query(max_length=100)] = None,
    emotion: Annotated[str | None, Query(max_length=100)] = None,
    person: Annotated[str | None, Query(max_length=255)] = None,
    place: Annotated[str | None, Query(max_length=255)] = None,
    is_archived: bool | None = None,
    created_from: datetime | None = None,
    created_to: datetime | None = None,
    page: Annotated[int, Query(ge=1)] = 1,
    page_size: Annotated[int, Query(ge=1, le=100)] = 20,
):
    logger.info(
        "Recall request received: has_query=%s query_length=%d has_tag=%s has_book=%s "
        "has_theme=%s has_emotion=%s has_person=%s has_place=%s thought_type=%s "
        "source_type=%s is_archived=%s page=%d page_size=%d",
        bool(q and q.strip()),
        len(q.strip()) if q else 0,
        bool(tag and tag.strip()),
        bool(book and book.strip()),
        bool(theme and theme.strip()),
        bool(emotion and emotion.strip()),
        bool(person and person.strip()),
        bool(place and place.strip()),
        thought_type.value if thought_type else None,
        source_type.value if source_type else None,
        is_archived,
        page,
        page_size,
    )
    query_embedding = None
    if q and q.strip() and settings.openai_api_key:
        try:
            query_embedding = OpenAIProvider().embed([q.strip()])[0]
        except Exception as error:
            logger.warning(
                "Recall semantic search unavailable: error_type=%s",
                type(error).__name__,
            )
    result = list_thoughts(
        db,
        user,
        RecallQuery(
            query=q,
            thought_type=thought_type,
            source_type=source_type,
            book_id=book_id,
            tag=tag,
            book=book,
            theme=theme,
            emotion=emotion,
            person=person,
            place=place,
            is_archived=is_archived,
            created_from=created_from,
            created_to=created_to,
            page=page,
            page_size=page_size,
        ),
        query_embedding=query_embedding,
    )
    response.headers["X-Total-Count"] = str(result.total)
    response.headers["X-Page"] = str(page)
    response.headers["X-Page-Size"] = str(page_size)
    response.headers["X-Total-Pages"] = str((result.total + page_size - 1) // page_size)
    logger.info(
        "Recall response returned: total_matches=%d returned_count=%d page=%d page_size=%d",
        result.total,
        len(result.items),
        page,
        page_size,
    )
    return result.items


@router.get("/deleted", response_model=list[ThoughtRead])
def list_deleted_thoughts_route(
    db: DbSession,
    user: CurrentUser,
):
    return list_deleted_thoughts(db, user)


@router.get("/{thought_id}", response_model=ThoughtRead)
def get_thought_route(
    thought_id: UUID,
    db: DbSession,
    user: CurrentUser,
):
    return get_thought(db, user, thought_id)


@router.patch("/{thought_id}", response_model=ThoughtRead)
def update_thought_route(
    thought_id: UUID,
    payload: ThoughtUpdate,
    db: DbSession,
    user: CurrentUser,
):
    return update_thought(db, user, thought_id, payload)


@router.delete("/{thought_id}", status_code=status.HTTP_204_NO_CONTENT)
def delete_thought_route(
    thought_id: UUID,
    db: DbSession,
    user: CurrentUser,
):
    soft_delete_thought(db, user, thought_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post("/{thought_id}/organize", response_model=ThoughtRead)
def retry_ai_processing_route(
    thought_id: UUID,
    db: DbSession,
    user: CurrentUser,
    response: Response,
    idempotency_key: IdempotencyKey = None,
):
    return execute_idempotent(
        db, user, f"POST /thoughts/{thought_id}/organize", idempotency_key, {}, response,
        lambda operation: retry_ai_processing(db, user, thought_id, operation=operation),
    )


@router.post("/{thought_id}/restore", response_model=ThoughtRead)
def restore_thought_route(
    thought_id: UUID,
    db: DbSession,
    user: CurrentUser,
):
    return restore_thought(db, user, thought_id)