# openapi.json
{
  "components": {
    "schemas": {
...
      "AIProcessingStatus": {
        "enum": [
          "not_requested",
          "pending",
          "processing",
          "ready",
          "failed"
        ],
        "title": "AIProcessingStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceType": {
        "enum": [
          "manual",
          "book",
          "article",
          "website",
          "audio",
          "import",
          "unknown"
        ],
        "title": "SourceType",
        "type": "string"
      },
      "StorageScope": {
        "enum": [
          "cloud",
          "local_device"
        ],
        "title": "StorageScope",
        "type": "string"
      },
      "ThoughtCreate": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "book_author": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Author"
          },
          "book_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "book_title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Title"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "manual_tags": {
            "items": {
              "type": "string"
            },
            "title": "Manual Tags",
            "type": "array"
          },
          "page_reference": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Reference"
          },
          "source_author": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Author"
          },
          "source_title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Title"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType",
            "default": "manual"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "storage_scope": {
            "$ref": "#/components/schemas/StorageScope",
            "default": "cloud"
          },
          "thought_type": {
            "$ref": "#/components/schemas/ThoughtType",
            "default": "thought"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "use_with_ask_my_mind": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use With Ask My Mind"
          }
        },
        "required": [
          "body"
        ],
        "title": "ThoughtCreate",
        "type": "object"
      },
      "ThoughtRead": {
        "properties": {
          "ai_processing_status": {
            "$ref": "#/components/schemas/AIProcessingStatus"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "book_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Author"
          },
          "book_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "book_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Title"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "manual_tags": {
            "items": {
              "type": "string"
            },
            "title": "Manual Tags",
            "type": "array"
          },
          "page_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Reference"
          },
          "purge_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purge At"
          },
          "source_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Author"
          },
          "source_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Title"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "storage_scope": {
            "$ref": "#/components/schemas/StorageScope"
          },
          "thought_type": {
            "$ref": "#/components/schemas/ThoughtType"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "use_with_ask_my_mind": {
            "title": "Use With Ask My Mind",
            "type": "boolean"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "title",
          "body",
          "thought_type",
          "source_type",
          "source_title",
          "source_author",
          "source_url",
          "book_id",
          "book_title",
          "book_author",
          "page_reference",
          "manual_tags",
          "storage_scope",
          "use_with_ask_my_mind",
          "ai_processing_status",
          "is_archived",
          "created_at",
          "updated_at",
          "deleted_at",
          "purge_at"
        ],
        "title": "ThoughtRead",
        "type": "object"
      },
      "ThoughtType": {
        "enum": [
          "thought",
          "journal",
          "quote",
          "book_excerpt"
        ],
        "title": "ThoughtType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Mind Palace App",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/thoughts": {
      "get": {
        "operationId": "list_thoughts_route_thoughts_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "thought_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ThoughtType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Thought Type"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "book_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book Id"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "book",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book"
            }
          },
          {
            "in": "query",
            "name": "theme",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          },
          {
            "in": "query",
            "name": "emotion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Emotion"
            }
          },
          {
            "in": "query",
            "name": "person",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person"
            }
          },
          {
            "in": "query",
            "name": "place",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place"
            }
          },
          {
            "in": "query",
            "name": "is_archived",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Archived"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThoughtRead"
                  },
                  "title": "Response List Thoughts Route Thoughts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Thoughts Route",
        "tags": [
          "thoughts"
        ]
      },
      "post": {
        "operationId": "create_thought_route_thoughts_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9_-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Thought Route",
        "tags": [
          "thoughts"
        ]
      }
    },
    "/thoughts/deleted": {
      "get": {
        "operationId": "list_deleted_thoughts_route_thoughts_deleted_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThoughtRead"
                  },
                  "title": "Response List Deleted Thoughts Route Thoughts Deleted Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Deleted Thoughts Route",
        "tags": [
          "thoughts"
        ]
      }
    },
    "/thoughts/{thought_id}": {
      "delete": {
        "operationId": "delete_thought_route_thoughts__thought_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "thought_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thought Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Thought Route",
        "tags": [
          "thoughts"
        ]
      },
      "get": {
        "operationId": "get_thought_route_thoughts__thought_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thought_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thought Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Thought Route",
        "tags": [
          "thoughts"
        ]
      },
      "patch": {
        "operationId": "update_thought_route_thoughts__thought_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "thought_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thought Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Thought Route",
        "tags": [
          "thoughts"
        ]
      }
    },
    "/thoughts/{thought_id}/organize": {
      "post": {
        "operationId": "retry_ai_processing_route_thoughts__thought_id__organize_post",
        "parameters": [
          {
            "in": "path",
            "name": "thought_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thought Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9_-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Ai Processing Route",
        "tags": [
          "thoughts"
        ]
      }
    },
    "/thoughts/{thought_id}/restore": {
      "post": {
        "operationId": "restore_thought_route_thoughts__thought_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "thought_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thought Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Neon-Auth-Session-Cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Neon-Auth-Session-Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore Thought Route",
        "tags": [
          "thoughts"
        ]
      }
    },
...
  }
}